licensing: Add X as a license to help licensing script.
libdrm comes from Xorg and inherits the MIT license but does not
provide any license in the sourc code.
I've added X to their license field so that the licensing script can
display the Xorg license for them too.
This is preferable from checking in an override that points their license
to a stale duplicate copy of the Xorg license.

BUG=chromium:323243

Change-Id: Id23dae658b3610a31b48e5ee64f0dcdaeeab1279
Reviewed-on: https://chromium-review.googlesource.com/192673
Tested-by: Marc MERLIN <merlin@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Marc MERLIN <merlin@chromium.org>