Revert 80MHz channels for some domains
Upstream enabled 80MHz on all regulatory domains in one fell swoop.
While it's  unclear in official communications from the various
regulatory bodies, it's becoming clear from multiple independent
sources that this may not be the case in some countries.  Add a
patch to upstream that will allow us to retract the 40 -> 80
bump for domains for which it is reported this is not allowed
yet.

BUG=chrome-os-partner:26747
TEST=regdbdump on re-flashed device

Change-Id: Ie80c815481ecbdf4639120ed6513d899f6ffd627
Reviewed-on: https://chromium-review.googlesource.com/193154
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>