libchrome: Add a library of libchrome dbus mocks
This allows writing unittests against the mocks defined in Chrome's dbus
library.  To do this, we add a header that allows us to include Chrome
mocks so that they look for gmock as part of our standard system
includes, and compile the mock classes into a static archive that we
can link unittests against.  Static archives get filtered out of images
that we build, so this adds no storage overhead.

TEST=Unit tests consuming Chrome mock classes now compile
BUG=chromium:360835

Change-Id: I6554986a6321db966fa0d759cc9b31e96ef394e4
Reviewed-on: https://chromium-review.googlesource.com/193650
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>