[chromiumos-overlay] Package server-side tests.
This brings back the packaging of server-side tests by creating an
ebuild for them.  This is done as server-side tests don't have deps, so
we don't need to force devs to go through the same long dep building
process as we do for client-side tests.

We also need to remove  the attempted `setup()` running/dep building for
server-side tests, as they don't have deps.

BUG=chromium:217273
TEST=equery files shows server-side tests still, bvt runs same set of
     tests

Change-Id: If677c6d97528abeba09f62f75729bd925d196727
Reviewed-on: https://chromium-review.googlesource.com/190422
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Gaurav Shah <gauravsh@chromium.org>