Chameleon: Remove the autotest-testdata-display ebuild
The test data, which currently contains EDID files, for the Chameleon test
display_Resolution, was compressed as a zipped tar file and put in GS server.
This ebuild is to install the test data to the test directory while running
build_package.

The lab Autotest server does not emerge this ebuild and the test data is not
installed on lab run. So we are going to move the test data to the test.

BUG=chromium:359483
TEST=build_packages
CQ-DEPEND=Iee79c9c05f5983d58b6f958fbdc951a934362a59

Change-Id: I6932228a986b8e92db1b0c0ac00bd81699212fca
Reviewed-on: https://chromium-review.googlesource.com/193583
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>