cros-kernel: disable multilib checking in /lib/modules
The multilib strict logic rejects the vdso objects we try to install
into /lib/modules when building on ABI=x32.  This can impact all
multilib systems, so add a general filter on the /lib/modules dir.
It's not really useful to scan it anyways.

BUG=chromium:220119
TEST=`emerge-x32-generic chromeos-kernel` no longer fails

Change-Id: I626d2ca1d39be5c2fde4fc42ac702862c997219a
Reviewed-on: https://chromium-review.googlesource.com/193552
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>