Remove autotest-tests dependency from autotest-chrome.
In https://gerrit.chromium.org/gerrit/39227, we added a 1.5 min build time
regression by making autotest-chrome (which takes 1.5 min to build) depend
on autotest-tests which tends to be the last thing built to handle merge
order. Because of this, all builds are 1.5 min slower.

BUG=None
TEST=build_packages

Change-Id: I4bb018365aeb826e8ce932f82a596660d8ebab66
Reviewed-on: https://chromium-review.googlesource.com/195252
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>