chromeos-base: drop munging of /etc/shadow for chronos
We want to delay the /etc/shadow creation to build_image, so let's drop
this logic from the ebuild.  It arguably shouldn't be here either.

By dropping this logic, we also can drop the last use of the old useradd
eclass, so let's clean that up too.

BUG=chromium:361864
TEST=`cbuildbot x86-generic-full` passes

Change-Id: I8676c8661520363343b12d8827f081339ba9b858
Reviewed-on: https://chromium-review.googlesource.com/196482
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>