xorg-server: drop x11-apps/xkbcomp
This tool is used to compile keyboard maps, but we don't do that at
runtime in the image (as all our maps are already compiled).  Drop
it from the X dep tree.  It's still in the sdk so we can compile maps
for the target at build time.

BUG=chromium:364168
TEST=`cbuildbot {arm,amd64,x86}-generic-full` passes

Change-Id: I6dbe2502d172d3ede9b45c9b5c9174cde2b2737d
Reviewed-on: https://chromium-review.googlesource.com/195419
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>