cros-kernel2: Make sure debug info / frame pointers are on for kgdb
When using kgdb we need debug info.  We've got that turned on for our
local kernel but don't have it on for all of our fallback kernel
configs.  One easy way to make sure it's always on for kgdb is to
specify it whenever kgdb is enabled.

BUG=chromium:365011
TEST=Set a breakpoint on snow and continue--it hits now.

Change-Id: I359062aad810bdbdcd90427f6efa214ad933e0d6
Reviewed-on: https://chromium-review.googlesource.com/196376
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>