Fix X11 dependencies for Salsa
Since the updated try_touch_experiment is including the X11 libraries to
alter the X input parameters directly now, it needs to depend on the x
libraries.  This CL adds the dependency.

BUG=chromium:364231
TEST=emerged for a clean/empty build directory and it pulled in all the
libraries it needed and built correctly.  Before this patch it was
unable to find a couple .h headers when it tried to compile unless the
other packages had been emerged previously by hand.

Change-Id: Ieb62160452e59e4fb064963658562f50fe6e84b3
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/196276
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>