Add lmbench ebuild
This is based on a version that was removed from Gentoo a few years
ago, but is somewhat simplified.  I updated to latest version that I
could find, and have to fetch it from sourceforge.  I also removed a
bunch of stuff from the old ebuild related to the Beaver Challenge
2004 and some other stuff that we don't care about.

BUG=chromium:354161
TEST=`emerge-daisy lmbench`
TEST=`emerge-link lmbench`
TEST=`emerge-x86-generic lmbench`

Change-Id: Ic342752e1bf73d75c5805cb38046beb424933516
Reviewed-on: https://chromium-review.googlesource.com/196348
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>