Remove chromeos-base/tty1 from chromeos-interface
tty1 was installed by chromeos-base/tty1 when the chromeless_tty use flag was
set. chromeos-base/tty1 is getting deprecated. This removes the dependencies on
it.

As the boards using chromeless_tty probably did not intentionally wanted tty1,
we do not install it by default.

BUG=chromium:361736
TEST=Build storm. tty1.conf is not installed.
TEST=trybot on mario, daisy, duck and link

Change-Id: Id9a7de1f9a38d711f1a42d126d15a3713bafd3b1
Reviewed-on: https://chromium-review.googlesource.com/196640
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>