cros-kernel2: Turn the watchdog subsystem off when enabling kdb/kgdb
kdb / kgdb doesn't know anything about the hardware watchdog subsystem
(though it knows about the NMI watchdog, softlockups, etc).  Rather
than a big overhaul to hardware watchdog, it's significantly simpler
to just disable the watchdog subsystem when we enable kdb / kgdb.  The
hardware watchdog is really just a pain for developers anyway since it
doesn't give you a chance to inspect the system--it just reboots it!

Alternative to this is a series like:
* https://chromium-review.googlesource.com/196650
* https://chromium-review.googlesource.com/196651
* https://chromium-review.googlesource.com/196652
* https://chromium-review.googlesource.com/196653

BUG=chromium:220761
BUG=chromium:365011
TEST=Drop into kgdb and don't see a watchdog reset.

Change-Id: I146c3c24d916ce18b16a3f0ef1d3eaccf7612bd8
Reviewed-on: https://chromium-review.googlesource.com/196642
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>