add src_test hooks for asan failures
As we enable ASAN in unittests more, we want to be able to properly log
and display the errors when they happen.  Currently, ASAN just reports
that it couldn't write to the default log path in /var/log (since it
doesn't have write permission).

BUG=chromium:367879
TEST=added an asan failure to bootstat and checked that it was caught & logged
TEST=`cbuildbot amd64-generic-asan` passes
TEST=`cbuildbot x86-generic-full` passes

Change-Id: Ib0ef6eeb6885066c83413d509592132a77bb0ae5
Reviewed-on: https://chromium-review.googlesource.com/197321
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>