Migrate tcpdump, nfqueue and dhcp out of chromeos-base
The tcpdump user is already handled in the (upstream) tcpdump
ebuild, so no action apart from removing it from chromeos-base
is necesary.

The nfqueue user is added to the pkg_preinst from platform2,
conditional on installation of shill

A pkg_setup rule is added to the dhcp ebuild to create the dhcp
user and group.

BUG=chromium:370132
TEST=emerge each of these packages

Change-Id: I47162ee974ae9e3842f58086f49cd5b2c5eb8bae
Reviewed-on: https://chromium-review.googlesource.com/198229
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>