Move devbroker and devbroker-access to permission_broker ebuild
Instead of managing these users and groups in the chromeos-base ebuild,
do it in the package that actually uses them.

Also, move the 'serial' group to pkg_preinst(). It's referred to
in udev rules in addition to being used by the permission_broker
so we should install it somewhere central-ish.

BUG=chromium:360813
TEST=build an image, check for these users and groups.

Change-Id: I75bc3433dbe5272a50f48b5d0d4a81f57122cd43
Reviewed-on: https://chromium-review.googlesource.com/198312
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>