make.conf: start unifying common settings
This introduces the framework for sharing settings between the sdk
and target boards.  The common-target file was supposed to be this,
but it grew into only supporting cross-compiled boards.

For now, we just lay the groundwork without moving any settings.
This way we can opt in new systems to it before we cut over the
actual settings and possibly breaking people.

It looks like:
 all sdk specific make.conf files inherit
   make.conf.common-sdk inherits
     make.conf.common
 all target specific make.conf files inherit
   make.conf.common-target inherits
     make.conf.common

BUG=chromium:371532
TEST=`cbuildbot chromiumos-sdk` still works

Change-Id: I40ae945aa5fd57adb1d3fdbda62365a09399242e
Reviewed-on: https://chromium-review.googlesource.com/201402
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>