chromeos-auth-config: fix handling of baked-in and devmode passwords
If a password is set in the root filesystem, the current auth file does
not play well with it -- if no devmode password is set, then the password
is simply ignored.

Add a check at runtime so that we do not allow passwordless login if the
current user has a password set in /etc/shadow already.

BUG=chromium:367909
TEST=`cbuildbot x86-generic-full` passes
TEST=no pass in /etc/shadow & no devmode pass -> log in w/out any pass
TEST=   pass in /etc/shadow & no devmode pass -> shadow pass required
TEST=   pass in /etc/shadow &    devmode pass -> shadow pass or devmode pass required
TEST=no pass in /etc/shadow &    devmode pass -> devmode pass required

Change-Id: Ia51381658edceb8265ab1c173ebc89d0ab6d5975
Reviewed-on: https://chromium-review.googlesource.com/200556
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>