sys-boot/*: Remove fwserial USE flag, always build serial and GDB images
This patch changes the way firmware images are built: while we
previously used the fwserial USE flag to turn on the serial console, the
new mechanism will always generate images both with and without serial
output (image.bin and image.serial.bin). In addition, it will build a
new image.dev.bin which includes both serial console and GDB support
(including the hack to trap into GDB through the recovery key
combination), which is intended as the new default image for firmware
and kernel developers (or really anyone who might be interested in
remote-controlling his firmware over Servo from time to time).

The reason for this (as opposed to adding another USE flag) is that this
is much easier to use for external vendors and people who are unfamiliar
with the firmware build process. Developers don't need to know how to
build a firmware image or how to pass USE flags... they can simply take
the right file they get from build_packages (or even from an official
firmware_from_source.tbz buildbot tarball) and flash it to the device.

As rsync and other ebuild thumb-twiddling seem to be the biggest time
sinks in our build system, the effective build time increase (for
essentially building everything twice) only seems to be about 20% for a
full emerge of all five ebuilds. I think this is a reasonable trade-off
for the increased convenience, especially since there are much more
pointless things we are wasting build-time with right now (I use scripts
for day-to-day development that can usually build a new image in ~10s).

CQ-DEPEND=CL:202536
BUG=chrome-os-partner:18390
TEST=Emerged everything on Nyan_Big and Falco, confirmed that all output
images do what they're supposed to.

Change-Id: Ie0554659ecc154b3c124189d232cce1c44ba7b7d
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202564
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>