cros-kernel2: raise size limit on 3.14+ kernels
8MB limit is really only there for recovery and factory images, and
we're not expecting to use 3.14+ kernels on the systems that have that
limitation. Lift it for newer kernels with that in mind.

The only overlay with 3.14+ and platforms that have the limit is for
booting generic overlays on them, and we rarely or never do that with
a recovery image, just with developer image + chromeos-install. So this
should be fine.

BUG=chromium:380848
TEST=USE="fbconsole initramfs tpm i2cdev vfat" emerge-amd64-generic chromeos-kernel-3_14

Change-Id: I6a33101b805e55de8856b47b6f8a83a18526e3ea
Reviewed-on: https://chromium-review.googlesource.com/202202
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>