Remove chromeos-base/verity from update_engine ebuild.
The update process doesn't use the verity tools to generate nor
veirfy the hashes. The generation is done on the /postinst file
from the new image and verification of the hashes is done by the
kernel on the next boot, as usual.

BUG=chromium:380893
TEST=emerge-panther_embedded --unmerge verity ; emerge-panther_embedded update-engine doesn't re-install verity
CQ-DEPEND=CL:202634

Change-Id: I665ca000bbe657f63b8ab61751209719136ea71a
Reviewed-on: https://chromium-review.googlesource.com/202617
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David Zeuthen <zeuthen@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>