chromeos-ec: Allow building host-side tools for board targets
Previously, to build host-side tools, the USE flag 'cros_host' needs to
be set. However, this flag cannot be set when building for board
targets. To solve this, let's add a new USE flag 'utils'. With this, one
can build host-side tools on the host side by:

  sudo emerge chromeos-ec

and on board targets by:

  USE=utils emerge-<board> chromeos-ec

or by adding 'utils' to package.use.

BUG=chromium:380222
TEST=emerge-beaglebone_servo chromeos-ec with USE=utils

Change-Id: Id683900aa7c0cd70933e88d15cbdea66c43e6771
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202829
Reviewed-by: Mike Frysinger <vapier@chromium.org>