autotest: Refactor client tests into separate ebuilds and behind USE flags
For Brillo boards, this lets us only build tests that are appropriate
for that board (e.g. no tests requiring Chrome if there's no Chrome).

In addition, separate autotest packages can be built in parallel, reducing the
amount of time it takes to build the entire set of client tests.

Finally, also removed some tests which do not exist in the source. These are:
 - login_DBusCalls
 - login_SecondFactor
 - platform_Shutdown

Unused tests with sources will be removed in a follow up change.

BUG=chromium:356414
TEST=trybot on x86-alex, daisy-release, lumpy-release, \
       lumpy-incremental, pre-cq-group, gizmo

Change-Id: If64e62951993f0a0e03890ac7d6084e6a53b8182
Reviewed-on: https://chromium-review.googlesource.com/203015
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Gaurav Shah <gauravsh@chromium.org>