Enable test to collect system tray images.
We are collecting system tray images in upcoming builds so as to have
data on 'default' state of the tray in the lab.

We do this so that we can establish some sort of a reference image.
Having a reference image will allow us to write image-comparison-based tests
against this reference.

BUG=chromium:374981
TEST=None. The lab has to run this one.

Change-Id: Ib41aa85238ce241ad5e646ace5c461565d4a8115
Reviewed-on: https://chromium-review.googlesource.com/203091
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>