perf: Get correct kallsyms start address
Applying a patch to the upstream perf sources:

=======================================================================
The function machine__get_kernel_start_addr() was taking the first
symbol of kallsyms as the start address. This is incorrect in certain
cases where the first symbol is something at 0, while the actual kernel
functions begin at a later point (e.g. 0x80200000).

This patch fixes machine__get_kernel_start_addr() to search for the
symbol "_text" or "_stext", which marks the beginning of kernel mapping.
This was already being done in machine__create_kernel_maps(). Thus, this
patch is just a refactor, to move that code into
machine__get_kernel_start_addr().
=======================================================================

BUG=chromium:372024
TEST=build perf successfully, collect profile, view in "perf report -D",
make sure kallsyms mapping does not start at 0.

Change-Id: If4781b421e5aed9c9ff196a5d36885e334b42b51
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203196
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>