mem.py: Lower mapping granularity to 4K pages
mem.py currently always aligns addresses to 64K when mmap'ing /dev/mem.
This will generally map more than necessary, and leads to problems when
the accessed page itself is accessible but something else in the same
aligned 64K region is not (e.g. because it has been marked as reserved
memory by firmware).

This patch reduces the alignment to 4K, which is the normal page size on
all our systems and thus sufficient for a successful mmap.

BUG=None
TEST=Manual

Change-Id: I03c84bd742e2528565d92363b78ffa5abc68791f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203315
Reviewed-by: Doug Anderson <dianders@chromium.org>