chromeos-factory: Remove unneeded dependencies
The packages we need for factory test are all included in test images.
There is no need to keep the dependencies and they are just making it
more difficult to generate factory toolkit. Currently we don't emerge
dependencies when building factory toolkit. (Otherwise it would pull in
a ton of packages, which are all already in the test image.) This,
however, would be harder if we need to make chromeos-factory a virtual
package someday. With virtual packages, we will absolutely need to
include dependencies, and the current factory toolkit generation process
won't work.

This also removes the currently unused memento_softwareupdate package
from factory image. If we need it in the future, we'll add it to the
test image.

BUG=None
TEST=Build a new factory image and check the packages involved are
included in the image.

Change-Id: I264806e7c0ff2fb66a23e10055b537faf0c671ec
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203622
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>