rsyslog: Remove redundant static library from imuxsock plugin.
imuxsock.c plugin uses some functions from sd-daemon.h that are
compiled as part of librsyslog.a. Including $(RSRT_LIB) to the LIBADD
variable on these plugins means that we are linking librsyslog.a
statically into that plugin. Nevertheless, the rsyslogd binary
already includes most of the librsyslog.a and exposes its symbols.
Because of this, we can let those functions undefined in the plugin
.so file and let dlopen() link those undefined functions against the
global symbols from rsyslogd.

This patch removes the static library from imuxsock.so and ensures
that all the objects from the librsyslog.a are included on rsyslogd.

BUG=chromium:382072
TEST=emerge-panther_embedded rsyslog; installed imuxsock.so is small.

Change-Id: Ic7d93d2330ba60548c435c629028f3ff10da7c07
Reviewed-on: https://chromium-review.googlesource.com/204432
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>