Disable C++ exceptions by default on all the ebuilds.
This patch adds a bash function "cros_enable_cxx_exceptions" that can
be called from chromiumos-overlay/chromeos/env/$DIR/$PKG to re-enable
the C++ exceptions.

This should be enabled in all the packages that explicitly use try,
catch or throw; and also on all the libraries that get involved in
passing the exception through a call. For example, if your
application calls a library with a callback pointer that throws an
exception when is called and you intend to catch it, that library
should also be compiled with exception support. If your application
uses exceptions but they don't involve a call to another utility
library, then that's not needed.

BUG=chromium:386443
TEST=cbuildbot {x86,amd64,arm}-generic-full
CQ-DEPEND=CL:205686

Change-Id: I4975607207be3934bb909d7940418570ff6b1328
Reviewed-on: https://chromium-review.googlesource.com/204622
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>