sys-apps/coreutils: Add --enable-single-binary option.
This patch pulls a reduced version of the single-binary (or multicall
binary) discussed on the coreutils mailing list, implemented by us.
The patch adds a new --enable-single-binary option to the configure
script that builds a single binary "coreutils" that uses argv[0] to
determine the utility being called. It also allows you to install
shebangs (#!) files instead of symlinks, allowing to create symlinks
to them installed tools and not relying in argv[0] for detecting the
selected tool.

This ebuild installs shebangs to the single "coreutils" binary
to allow other programs to call the tools using symlinks with
different names.

Finally, due to some changes on the manpages generation, we also
patch the files to provide the coreutils.1 man page and avoid
re-building it when installing since that's not possible on
cross-compiling environments.

This options is enabled with the multicall USE flag.

BUG=chromium:383000
TEST=USE=multicall emerge-link coreutils; verified installed files.
TEST=USE=multicall sudo emerge coreutils; verified installed files.
TEST=cbuilbot chromiumos-sdk gizmo-release

Change-Id: I8b458e869a271add2fa6d2c190677a042833bc08
Reviewed-on: https://chromium-review.googlesource.com/206396
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>