mod_wsgi: add fix for apache2.4.x threads crashing.
We noticed that child threads in apache tend to crash a lot and
when running tests on moblab, crashes were observed and reported.

This brings in a fix from
http://code.google.com/p/modwsgi/source/detail?path=/mod_wsgi.c&name=mod_wsgi-3.X&r=bdbeacb88f
to address this problem.

BUG=chromium:388519
TEST=apache is no longer acting like a drunk driver and crashing
all over the place. test_that completes tests without crash collecting.

Change-Id: If08192556321fb0f4f431217d598bdcb0b1b8349
Reviewed-on: https://chromium-review.googlesource.com/205684
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>