libva-intel-driver to use newer version of gpu-tools
When bumping up the version of intel-gpu-tools to 1.7 then the
intel-gen4asm tool is used to generate the shader bits.

The proper location of intel-gen4asm is not accessible through
the PATH variable when cross compiling so it should not be used.

The patches added can be dropped when bumping up the version of
this ebuild.

Bug=None
Test="emerge-${board} intel-gpu-tools-1.7; emerge-${board}
libva-intel-driver-1.3.1; where ${board} is every x86 and x86_64 target
available with its variants"

Change-Id: I85f8355f6e609751a7031b77c08b4d9664f00ca1
Signed-off-by: Daniel Charles <daniel.charles@intel.com>
Signed-off-by: James Ausmus <james.ausmus@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/206314
Reviewed-by: Mike Frysinger <vapier@chromium.org>