cros-kernel2: Add an rockchip fallback defconfig
This rockchip_defconfig is based on a something halfway between what
Heiko had in
<https://github.com/mmind/linux-rockchip/tree/topgit/rk3288> for
rk3288_defconfig and what we have for exynos5_defconfig.  I have
verified that it boots to serial command prompt on a kernel based atop
Heiko's tree.

While this is not a true upstream fallback config yet, it's pretty
close.

BUG=chrome-os-partner:30167
TEST=See this config used to build; boot the kernel

Change-Id: Ia3b3a5fa4fa68003cc9e7f4471c759a335af3b00
Reviewed-on: https://chromium-review.googlesource.com/206374
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>