Build "mini" PAR to run a subset of binaries on test images.
chromeos-factory-mini is a subset of the factory software that can be
used to run utilities like gooftool, hwid, and regcode, which may be
useful in the CrOS test environment.  For instance, this would allow
"gooftool probe" to be used to probe hardware components in Moblab.

We don't want to install the entire chromeos-factory package in the
test image, since it is quite large, so this package comprises a small
".par" file (/usr/local/factory-mini/factory-mini.par) containing the
necessary subset of factory Python code, and symlinks from
/usr/local/bin to that file.

BUG=chrome-os-partner:30468
TEST=cbuildbot
CQ-DEPEND=CL:207488

Change-Id: I325ffb0bf825300949bee8dd0831504f08bd781e
Reviewed-on: https://chromium-review.googlesource.com/207496
Tested-by: Jon Salz <jsalz@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Jon Salz <jsalz@chromium.org>