autotest-server: Make power_management USE flag default on
We never explicitly set the USE flag but do disable it for some
boards. So have to ebuild assume it to be default on if not set.

Also, add a missing '?' in the USE flag conditional for chromeless_tty.
This was resulting in "QA notice" warning and the USE flag conditional
being ignored.

BUG=chromium:394112
TEST=cros_workon --board=panther start autotest-server-tests;
  emerge-panther autotest-server-tests shows the power tests
  as being installed in the sysroot.

Change-Id: I7e3ae468df8cdc706df122affc4fa239e3189e0f
Reviewed-on: https://chromium-review.googlesource.com/208200
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Gaurav Shah <gauravsh@chromium.org>