update_engine: Move sys-fs/e2fsprogs build time only dependency.
Dependency on ext2fs.h was moved to delta_generator and unittest code
only. This means that sys-fs/e2fsprogs is a build time dependency
because it is needed for unittest and a runtime dependency when the
delta_generator USE flag is set. This patch reflects that change in
the ebuild.

BUG=chromium:394184
TEST=FEATURES="test" emerge-link update_engine; sudo emerge update_engine
CQ-DEPEND=CL:If750e4834d42bbeaeddc6eff7c4690dd916c73bf

Change-Id: I373ddd8a5d6b184ce5f56800050f82bd299f5dc4
Reviewed-on: https://chromium-review.googlesource.com/209003
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>