sys-boot: Adapt to new developer image mechanism
This patch adapts the depthcharge and chromeos-bootimage ebuilds to the
new depthcharge Makefile. We explicitly link the three depthcharge
builds we need for the current board, and reuse the individual object
files so we don't have to run the whole build twice anymore. The
image.dev.bin build is switched to pick up the new dev.elf binary from
the normal depthcharge output directory, and is prepared for netboot in
the same way as image.net.bin.

CQ-DEPEND=CL:209273
BUG=chrome-os-partner:18390
TEST=Built and booted image.bin, image.net.bin and image.dev.bin on
Daisy, Nyan_Blaze and Falco.

Change-Id: I42abb61910e73e17bbe74e0e7a28daa9231d5a3d
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209260
Reviewed-by: Aaron Durbin <adurbin@chromium.org>