Prepare to upgrade libpng to 1.2.51
libpng 1.2.51 or later is required for FreeType 2.5.3 for bug 359004.
libpng 1.6.x (stable on some arch's) require a chain of packages to
be updated. For now, we're going to libpng 1.2.51 to quickly finish
upgrading FreeType to 2.5.3 (with some security fixes and much better
auto-hinting).

This CL does not upgrade libpng, yet (it's masked), but just adds a pristine
upstream ebuild and Manifest entry before making a few local changes.

BUG=chromium:359004
TEST=emerge-${BOARD} libpng still builds libpng-1.2.49

Change-Id: I6e1882244e488bd63869af3a5b15bba78b5b8c78
Reviewed-on: https://chromium-review.googlesource.com/209934
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>