Remove circular dependency between udev-208 and udev-171
As part of the transition to udev-208, the two versions of udev were
slotted and made dependent on each other to make sure that both were
installed.  Emerge currently issues a warning about this, and randomly
breaks one of the dependencies.  As virtual/udev currently depends on
>=udev-208, removing udev-171's dependency on udev-208 should still
allow both to be installed while getting rid of the warning.

BUG=chromium:397326
TEST=`./build_packages --board=panther` does not warn about circular
  dependencies; both udev-171 and udev-208 are emerged.

Change-Id: Iccb434084586ce6b3592fd677af786a672f05814
Reviewed-on: https://chromium-review.googlesource.com/210130
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Steve Fung <stevefung@chromium.org>
Tested-by: Steve Fung <stevefung@chromium.org>