glib: Install gresource only on the host.
gresource is a tool to list and extract binary resources included in
the ELF files using the GResource glib interface. This tool isn't
required to access these resources from the running ELF file, but
only to inspect it for debugging purposes.

This CL prevent's this tool from being installed on the image and
keeps it available on the host. It also removes the dependency on
elfutils from the devices since this is the only tool from glib
requiring it.

BUG=chromium:386488
TEST=`./build_image --board=panther_embedded` doesn't have this file.

Change-Id: Iaf4c6461890b9a5fd81090abe2ee2fdd1b3e1fb4
Reviewed-on: https://chromium-review.googlesource.com/210563
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>