x11-drivers/xf86-video-intel: Add kms-only USE flag
By default, xf86-video-intel builds in support for both ums and kms.
The vast majority of modern systems, however, only use kms not ums.
This in itself is not a huge deal, except UMS support also requires
DRI1 and GLX.  Thus, building ums support is not compatible with
an xorg-server that does not have DRI1 or GLX (ie, a server optimized
for use with EGL using only DRI2).

This patch adds a USE flag that allows specifying building kms only,
and further, enables it by default.  If a system still wants to build
both ums and kms it can by disabling this flag with USE="-kms-only"

BUG=chromium:340268
TEST=emerge-link_freon xf86-video-intel
TEST=emerge-link xf86-video-intel

Change-Id: I1623e97d76080313d0aa6923925035347de16b8a
Reviewed-on: https://chromium-review.googlesource.com/210940
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>