Fix errors in local.conf of fontconfig
The way to use <or> was incorrect in local.conf leading all the
fonts to lose subpixel rasterization.

Currently, there's no succinct way to apply edits to a set of font
families in fontconfig. (see
  https://bugs.freedesktop.org/show_bug.cgi?id=82582 and
  https://bugs.freedesktop.org/show_bug.cgi?id=33644 ).

I'm using the simpler of two methods suggested by Raimund Steger at

http://lists.freedesktop.org/archives/fontconfig/2014-August/005298.html

to check if font family is one of multiple names.

BUG=chromium:403502
TEST=After deploying the updated fontconfig on a device, run
`fc-match -v Arial | grep rgba` to get '1' instead of '5'.
TEST=run `fc-match -v MSung B5HK | egrep '(hint|alias|rgba
' to check the values are what's set for the font in local.conf
TEST=run `fc-match -v "Noto Sans Thai" | egrep '(hint|alias|rgba)'` to
check the values are what's set for the font in local.conf.

Change-Id: I21cc2fadf53bb06bec5de38595c94179a6397377
Reviewed-on: https://chromium-review.googlesource.com/212299
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Commit-Queue: Jungshik Shin <jshin@chromium.org>