chromeos-factory-deps: new package for listing factory software dependency.
To make it clear for what the factory software needs (and to have better
customization for each platform), we want to isolate the factory dependency from
test image.

Note chromeos-factory-deps is currently included if we are "not" building
embedded platforms. This is to prevent breaking or bloating embedded image
immediately. After the chromeos-factory-deps is tested and verified on embedded
images, we will move it to CROS_COMMON_RDEPEND.

BUG=chromium:403644
TEST=emerge-link chromeos-factory-deps
     build_packages; build_image test

Change-Id: Idd203cf37198b7ea7b23fa1d53ae0346c701663b
Reviewed-on: https://chromium-review.googlesource.com/212420
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>