sys-fs/avfs: Drop support for dynamic modules.
All the avfs modules are statically compiled in so no modules are
loaded dynamically on startup. This patch adds a flag to remove
the dynamic module loading support given that it is not used.

This helps keep track of the dependencies reducing the programs using
dynamic loading.

BUG=chromium:404692
TEST=`objdump -T avfsd | grep dl` doesn't show any dl* symbol used.
TEST=`equery-link f avfs | xargs -I {} objdump -T /build/link/{} | grep av_init_module` \
    didn't show any module before this patch.
TEST=cbuildbot x86-generic-full amd64-generic-full

Change-Id: I2f533610786e9e759610e2408f919f5f664d631c
Reviewed-on: https://chromium-review.googlesource.com/212922
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>